+++ Flashstudy_Mobile/app/src/main/java/br/com/flashstudy/flashstudy_mobile/offline/dao/UsuarioDao.java
@@ -31,5 +31,5 @@ public interface UsuarioDao {
     List<UsuarioOff> getUsuarios();
 
     @Query("SELECT * FROM usuario WHERE codigo = :codigo")
-    UsuarioOff getUsuarioById(int codigo);
+    UsuarioOff getUsuarioById(long codigo);
 }
